// frontend/turmerik-terminal/src/types/project-data.ts
export interface ProjectData {
    project_num: string
    project_title: string
    organization: string
    total_cost: number
    estimated_roi: number
    voi: number
    roi_per_dollar: number
    public_roi: number
    estimated_qalys: number
    funding_agency: string
    project_url: string
    uses_dct: boolean
    investment_decision: string
  }
  
  